Add more logic to latitude and longitude to allow more granular control on when we will poll data from Openweathermap
Change the type of latitude and longitude from string to Decimal in some places. Needs to be completed in another commit.

Adjust config_template.py to have a list of allowed coordinates where we will get weather data from Openweatherdata and the option to allow requests for all coordinates (if this is enable on a website accessible to anybody, there could be quote high costs for the API usage)

Add ability to not just update existing coordinates in sqlite.py but also add new ones.

Add logic to allow or disallow polling of new sensor data from openweathermap API depending on the requested coordinates. The coordinates will be a Decimal of up to 7 decimal places.